delete your acount, fix

+++ app/Http/Controllers/API/ProfileController.php
@@ -202,9 +202,15 @@ class ProfileController extends Controller
 
     public static function deleteYourAccount(Request $request)
     {
-        $comps = ChallengeCompletion::where('user_id', $request->user_id)->delete();
+        
         $user = User::find($request->user_id);
+        if(!$user) {
+            return response()->json([
+                'account not found' => 'account ' . $request->user_id . ' not found'
+            ]);
+        }
         \Log::info('USER DELETE: ' . $user);
+        $comps = ChallengeCompletion::where('user_id', $request->user_id)->delete();
         $user->delete();
         return response()->json([
             'deleted' => $user
